% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reading_nfocus_xls.R
\name{add_focus}
\alias{add_focus}
\title{Add column Indicating Rows With Focus Name}
\usage{
add_focus(df)
}
\arguments{
\item{a}{dataframe}
}
\description{
Add column Indicating Rows With Focus Name
}
\seealso{
add_gender add_states
}

