% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reading_nfocus_xls.R
\name{add_gender}
\alias{add_gender}
\title{Identifies Gender-related rows}
\usage{
add_gender(df)
}
\arguments{
\item{df}{a dataframe}
}
\description{
The nfocus datasets typically have several different demographic information
for which there are health statistics (e.g., race, gender, etc.). This
function can be used to identify which rows of the dataset contain data
by gender.
}
\details{
This function works by adding another column with \code{logical} values. A
value of \code{TRUE} indicates the row has statistics for by gender, and
\code{FALSE} signifies otherwise
}

