% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reading_nfocus_xls.R
\name{add_state_names}
\alias{add_state_names}
\title{Adds a Column With State Name}
\usage{
add_state_names(df)
}
\arguments{
\item{df}{a dataframe}
}
\description{
NB: assume that there is a State_row column present in the dataframe
}
\seealso{
add_focus_name
}

