% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xls_reading.R
\name{fix_col_names}
\alias{fix_col_names}
\title{Simplifies Working with columns in the xls files}
\usage{
fix_col_names(df)
}
\arguments{
\item{df}{}
}
\description{
Removes empty columns, and renames the columns with non-unique names.
}

