% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gender_comp.R
\name{get_gender_comp}
\alias{get_gender_comp}
\title{Obtain Health Conditions by Gender}
\usage{
get_gender_comp(fname)
}
\arguments{
\item{fname}{filename, as a string}
}
\description{
Obtain Health Conditions by Gender
}
\examples{
mental <- get_gender_comp("../../jhuDash_pkg/raw_data/NFOCUS03ST.XLS")
}

