% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xls_reading.R
\name{rm_tail}
\alias{rm_tail}
\title{Remove the last lines of an excel sheet.}
\usage{
rm_tail(df, num_rows = 6)
}
\arguments{
\item{df}{a dataframe}

\item{num_rows}{number of rows to remove from end}
}
\description{
This function aims to remove the documentation often contained included at
at the tail end of excel document. Those last lines also
tend to have hyphens in the "Objective" column, which interferes with the
heuristic used to generate a column for the name of the FOCUS that data
corresponds to.
}

