% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reading_nfocus_xls.R
\name{fix_col_names}
\alias{fix_col_names}
\title{Simplifies Working with columns in the xls files}
\usage{
fix_col_names(df)
}
\arguments{
\item{df}{}
}
\description{
Removes empty columns, and renames the columns that have non-unique names.
}
\details{
NB: Not sure what is represented by the values in the rows with state names.
 Here those values are denoted by "rounded."
}

